Add files via upload
Currently no colorbars on the graphs. Adding colorbars changes the aspect ratio of the graph which causes the images not to align properly. Currently planning on adding a separate div with just the colorbar. To view colorbars in this version uncomment lines 189-191, but this causes the images not to load properly